Allow flexibility in the first DICE cert config
Begin the introduction of a session context that is initially used for
passing around options. The first option allows the config descriptor of
the first DICE certificate to be ignored if it doesn't conform to the
Android specification.

This option is always enabled in the cxxbridge library used by VTS to
allow for ROM implementations that got it wrong for Android U, where it
isn't yet important.

The option can be enabled in the `hwtrust` tool by passing the new flag.

Bug: 261647022
Test: atest --host hwtrust_cli_tests hwtrust_tests libhwtrust_tests
Change-Id: Ic0bb3856fd4ce076f4b94bb4a80ae116faaed3d8