Add an action for invoking Android Power Menu
Starting with API 31 holding the Power button invokes Assistant instead of the Power menu. The Power menu is invoked by pressing Power and Volume Up together. The new action simulates pressing these two buttons.

By default, the Power+Volume Up action is not included in the Running Devices toolbar, but it can be invoked by hitting Shift Shift or by assigning a shortcut to it.

Test: included
Bug: N/A
Change-Id: If1f57289bc1bb6fd6ee43bcc5feb42607581a93d