Autoupdate of 2 submodules
*tools/adt/idea updated to 5b9b435e54e89b68a08dd49200d859e4e05b5cfe
Make app insights code sharable

This change follows up on the previous one and updates the files in place to make them usable. Some of these classes have been made generic in an attempt to make them reusable in the future when we integrate play vitals API.

They are:
1) AppInsightsState
2) AppInsightsModuleController

TimeIntervalFilter and Fatality types are moved to the new module. They are made to include all possible values used by both CL and Vitals.  Users can choose which values they want to use.

Test: n/a
Bug: 263259654
Change-Id: I48ba13ffc4d21af6cab46ee554d48e5dc13bd671

*tools/vendor/google updated to 08971ecd8b617da329d14277f0eb38e270f6e3ae
Update firebase to depend on app-quality-insights module

Also TimeIntervalFilter and Fatality types are refactored out of firebase and made sharable.

Test: n/a
Bug: 263259654
Change-Id: I2a92a2651f812e35cb6f0524aec328214df77c75