Autoupdate of 2 submodules
*tools/adt/idea updated to fa9736c22549b35d574d61870effe0f081062be4
Report `useAndroidX` from the build system models

... rather than letting the IDE read directly from the
`gradle.properties` file, to unblock changing the default in a future
version of AGP.

Includes a fallback injected model builder to ensure the value can
always be reported as part of the model, even when importing projects
using older AGPs.

Bug: 258855275
Test: Updated IDE model snapshot tests
Change-Id: Ifb6b574942beac7cd6bf1e976db7ec0998624e77

*tools/base updated to d0dada68124bd32999de92bfa57d99276400f5aa
Pass `android.useAndroidX` property's value to the IDE

The IDE needs to read the property's value from the model,
and not directly from the `gradle.properties` file.

(This is required to set `android.useAndroidX = true` by default
in an upcoming change.)

Bug: 258855275
Test: Updated model golden files
Change-Id: I9ae6268c6634c5b49292f50e0d8c3eebb78ec490