tfrecordmaker: utility to create tf.record from a folder
This CL adds a utility that generates a tf.Record from
a folder. Every file becomes a tf.Example with 2 features:
a "text" feature that contains the file's contents, and
a "path" feature that contents the relative path of the
file from the content root.

Only files with an allow listed set of extensions are
processed.

The content folder can optionally contain a "tf-excludes.txt"
file that can reject certain file paths using globs.

The Example.proto file includes a subset of the tf.Example
proto that is needed for our purposes. The TFRecordWriter
is an implementation based on the spec for TF.Record,
inspired from a Java implementation present in the
tensorflow-ecosystem github repository.

Bug: N/A
Test: added
Change-Id: I645ebbcef25f7bd706c0d305eadc6cbfb03d2832