Autoupdate of 2 submodules
*tools/adt/idea updated to 18cf458db6b57045c05bb0771cf7d37905db1d88
Make AndroidModuleInfo an interface

Rendering pipeline currently depends on AndroidModuleInfo. We are
trying to make rendering pipeline independent of studio and it
seems impossible to reuse current AndroidModuleInfo implementation
for it heavily depends on studio-specific object, such as
AndroidFacet, AndroidModel etc. Thus we make AndroidModuleInfo an
interface renaming the previous implementation to
StudioAndroidModuleInfo and allowing different implementations
to be plugged in to rendering pipeline through AndroidModuleInfo
interface.

Bug: 265765901
Test: all tests should pass
Change-Id: Ied34fd0782602298fe2e29d33e274adcfc642678

*tools/studio/google/cloud/testing updated to 79462b50e349ca4429cbd54c534dfe35002221fa
Follow the AndroidModuleInfo API change

Bug: 265765901
Test: all tests should pass
Change-Id: I681abb0b35353f24692b1246cb8757bf0b6b8e0b