Autoupdate of 2 submodules
*tools/adt/idea updated to dd9a06a68396341cc5f1b3a867a624cab8138a53
Add hinge config for resizable AVD

Resizable AVD can switch to foldable mode. We add the hinge config to
resizable so that it can be folded in foldable mode.

Bug: 266975599
Test: AvdManagerConnectionTest.testResizableAvd
Change-Id: If96e74e1f5b56b4fe6d5cc94733f43119a86d98d

*tools/base updated to 0c93fe437c2580acf9838f96c061557c7eb64c6f
Add folded area for resizable AVD

Resizable AVD can switch to foldable mode. We add the folded area so
that it can be folded in foldable mode.

Bug: 266975599
Test: AvdManagerConnectionTest.testResizableAvd
Change-Id: Id491305b5e183afd6c44666dbfe61c46e456e02e