Display feedback buttons and action chunks
It's still pretty crude in terms of how we display
actions and where, but it's good enough for now.
I'll be reworking how the messages are displayed
next and this contains important prep work for it.

Adding actions has not been thoroughly tested, but
should work well enough for now, until the UI is
changed and we can work on it.

Bug: 268612652
Test: manually tested
Change-Id: I9f5462f49f09b3aebd3227b4cc721b9e3e08e29a