Autoupdate of 4 submodules
*prebuilts/tools updated to 869f52541fc0808ddfe7b3ef130800ee1869ce88
Add org.java-websocket:Java-WebSocket:1.5.3

This CL was automatically generated by running
"tools/base/bazel/maven/maven_fetch.sh" after adding
"org.java-websocket:Java-WebSocket:1.5.3" to artifacts
in `tools/base/bazel/maven/artifacts.bzl`.

Bug: 268244519
Test: N/A
Change-Id: I5728cfca66548aa060acf7e64b18bcb5bdffad6a

*tools/adt/idea updated to 4a5c9f0ef83731d46ee32ea838e43ee5ec7a7872
Add modules for Studio ML plugin

Note that this CL only adds the modules to the list
of modules used in the default IntelliJ project, it doesn't
actually include the module in the AS distribution.

Bug: 268244519
Test: N/A
Change-Id: If224d6cd373e0a1487a6623700f1fceee882e8a4

*tools/base updated to aa2d1df3a2ee5375e8dc3148c7c262cff296d652
Add org.java-websocket:Java-WebSocket:1.5.3

The Studio Bot plugin has a temporary dependency on
websocket that is likely to be there for 2-3 months.
This CL adds a dependency to a simple websocket library.

Bug: 268244519
Test: N/A
Change-Id: Ic3736a832f63c4c93818d6f12aee2b0611e17348

*tools/vendor/google updated to d7b87516795fcebf0e09f32d7e292f5603338c18
Add BUILD files to enable presubmit testing

The .iml files are included in the tools/adt/idea/.idea project,
so the ml/sml plugin will show up when opening that project.

Note that this means that our project files are spread across
3 build systems: gradle, intellij and bazel.

Bug: 268244519
Test: N/A
Change-Id: I69ad4e75b9826ca90f69bf8db78a8f58b9dbb856