Reformat only: Switch lint's codebase formatter from ktlint to ktfmt
**THIS IS A FORMATTING CHANGE ONLY**: If you are using git annotate
to look up why a certain line changed, consider using IntelliJ's
"Git > Show History For Selection" feature after selecting a relevant
code fragment, which jumps across formatting barriers like this.

 Lint has been using an enforced code style for many years now. For
Java, it's using google-java-format, which is enforced with a
presubmit repo hook. For Kotlin however we were using ktlint, but
this was not enforced through repo upload hooks, so it was *very*
frequently forgotten, and our commit history is full of fixup CLs
touching unrelated files. Ktlint is also not as opinionated as ktfmt,
and has been changing quite a bit, so lint's current codebase is
formatted with a slightly older version of ktlint.

Ktfmt is now being used by many other subprojects in our codebase,
as well as Google overall, so this switches us over to the same
mechanism, as well as repo hooks.

Test: Existing
Bug: None
Change-Id: Ib958a4a08cd0f5eb94471083f4c4531571d6781b