sml plugin: add resource root
src/main/resources was not marked as a resource root in the iml,
which meant that this was not recognized as a plugin when launching
as part of the tools/adt/idea project. With this change, this
plugin gets loaded properly when launching the "Android Studio"
run configuration as long as this plugin is added as a dependency
to the `studio` module.

Bug: 268244519
Test: Launch Studio, confirm SML plugin is loaded.
Change-Id: I6b87cf3225dc113506fd1ee01d79b93ca0811c65