Simplify logic to start animation/interactive preview
We used to be able to start interactive preview from Animation Preview
and vice-versa, but it's been a while since that's not the case. We can
only start these modes from static preview. For that reason, we can do
a minor clean up to simplify the code to start these modes.

Bug: N/A
Test: N/A
Change-Id: I12817ebfaa44ebfe37c12a5ecb77932e76d4cea0