[AS-K2] Add K2 support to AndroidKtPsiUtils.
Add Analysis API support for utility functions in AndroidKtPsiUtils.
This should unbreak numerous K2 tests that are currently broken.

These functions will accept a KtAnalysisSession, if one is available, or
will call analyze() on their own if a session is not passed.

In addition, because the entire intellij.android.common test target is
now passing, promote that target to presubmit.

Bug: 269207454
Test: bazel test //tools/adt/idea/android-common:all
Change-Id: I8e951cb2b39c442d513ddac40401a2d6e1f125e1