Autoupdate of 4 submodules
*tools/adt/idea updated to 133ce3e9809a5b8dd39cecc2b226765592208d37
Create new interface ResourceRepositoryManager with only functions used.in rendering pipeline
And make StudioResourceRepositoryManager implement it

Currently we have extracted out only 3 functions that are used in the rendering pipeline. Everywhere else, we will continue using StudioResourceRepositoryManager. We will expand this interface as required in the future

Bug: 267600049
Test: StudioResourceRepositoryManager
Change-Id: I4b522f450039ecbd29cfa4928ca6dfce8b0a4411

*tools/studio/google/cloud/testing updated to a79a2d082a6825d1b6de2e2812401f8e01621ea6
Rename ResourceRepositoryManager -> StudioResourceRepositoryManager

Bug: 267600049
Test: unit test
Change-Id: Ibe07c0858107dbf145781b1c9629500293b206f3

*tools/vendor/google updated to f5779a980140787504ecd0dd4b8bc5ca3d54b002
Rename ResourceRepositoryManager -> StudioResourceRepositoryManager

Bug: 267600049
Test: unit test
Change-Id: Iee57a3a6ddf5e064f0c6b68095d180faa95c222f

*tools/vendor/google3/blaze updated to 2d0eb7c873496bb59b817bc9049a070e6d82abc7
Rename ResourceRepositoryManager -> StudioResourceRepositoryManager
Remove unused file

Rendering pipeline currently depends on ResourceRepositoryManager. In an effort to make rendering pipeline independent of studio, we are extracting an interface from ResourceRepositoryManager as it
seems impossible to reuse it because the current implementation
heavily depends on studio-specific classes like
AndroidFacet.

Bug: 267600049
Test: unit test
Change-Id: I7b5b4df59dacaa6983e0b908d42425608ca0cc18