Adjust the code snippet editor font to match message size
The default editor size is sometimes larger than the default font
size in the IDE. In that case, the code snippets look a little
out of place with the surrounding message text.

This CL tweaks the editor font size in code snippets to match
the surrounding text.

See https://screenshot.googleplex.com/BFzEuBFhUyMFFs3
for how this looks before and after for me.

Change-Id: I36ece8c7370fd8da4cbf2df2f6a2caaa7766bf18