Add actions to text messages
Currently, only code blocks can have associated actions. This CL also
adds relevant actions to text blocks, as well as an initial action: show
documentation. There is a topic recognizer which finds common topics
mentioned in the text, such as "baseline profiles", and adds an action
to show the relevant documentation.

Note: This adds the action list to the model; it isn't actually
displayed in the MessageChunkPanel yet.

Test: Unit test included
Change-Id: I90b98da386ea7fe1341e71d72b2d16de0b9a0ed4