Updating the results cache triggers a stateful button refresh
Due to background filters, buttons no longer update state properly. The
refresh they normally trigger after an action handler finishes happens
before the filter finishes and updates the cache so you don't see the
result of a click until _another_ click triggers the update.

To fix this, we have the results actor trigger an update when it writes
to the results cache.

Fixes: 271184757
Test: N/A
Change-Id: If14ad96400ab59c1c79c9f20fee41738cf97a5f1