Autoupdate of 2 submodules
*tools/adt/idea updated to 55da655966d5c98118d5c748d7613aab8362e24f
Allow customization of StatefulButtonMessage

As a data class, StatefulButtonMessage was effectively final. But it
was also a bit weird to be one, as a subclass of JPanel. This change
adds hashCode and equals only covering the properties that matter
for the data class aspect.

Bug: 268526991
Test: existing
Change-Id: I5bfed855299c2dab6dc17bd51ccaadf0e0a411cb

*tools/vendor/google updated to 5dbe5b60cabef1ea6eb0160fd08a71bd27daf63a
Navigate to code location from assistant action

We further extend StatefulButtonMessage to include custom UI, so
that we can override the hyperlink navigation and link to code.

Bug: 268526991
Test: existing
Change-Id: Ie9fdb5363e2bd176ebd3299e4b0bfcab3689fecb