adblib: Improve process properties collector
* In order to handle corner cases (timeout, process
termination, etc.) while collecting process properties,
the error handling logic has moved to the
`JdwpProcessPropertiesCollector` class.

* Add a `completed` property to `JdwpProcessProperties`
so callers know if a process has been fully retrieved.

* Improve EOF handling in case a process is terminated
while collecting its properties, i.e. do not log a `info`
message (since this is a common case).

* Stop collecting properties as soon as we know we
received everything we may need. This currently only
happens when the `WAIT` packet is sent, but there
will be more cases in the future.

* Do not send the `REAQ` packet, as 1) this is not useful
and 2) this is not supported in recent API level (30+).

* Add a bunch of unit tests to ensure code coverage
of this tricky area.

Bug: 188067646
Test: Included
Change-Id: I4486e0b2d7401594f1f9584be21f575948e873e8