Autoupdate of 2 submodules
*tools/adt/idea updated to 415c140b01f0c579645377080c31f8277a6fbcc4
Fix display of SDK ext in SDK Manager

by adding the extension level to the API column.

This change also removes the hiding of SDK extensions
when the details are not shown. There's no reason to hides
them as they are technically a different release.

Fixes: 270987669
Test: updated existing tests
Change-Id: Id51f5fdc4641791228b83e3d09fd8e8a8ef54661

*tools/base updated to 70521bcc81774d984a799e26ed7f861f7eb9e550
Deprecate AndroidVersion.getApiString

There are now 2 methods that explicitly declares
whether the string will contains the SDK extension value.

Bug: 270987669
Test: updated unit tests
Change-Id: Ife39f713a9df9a37d1a10a5f377caf001fc92bbe