Autoupdate of 2 submodules
*tools/adt/idea updated to fbf0492c8b81ab9afbfea51ba89d96185e46e2ce
Move desugarLibraryConfigFiles to the module system

They are only reported in the models of Gradle
projects that have the feature enabled, and
therfore there isn't a single answer for
the intellij project.

The context that needs this information has a
module anyway, so it seems fine to just move
it to the Module system from the project system.

Bug: 262468597
Test: No semantic change for existing tests.
Change-Id: Icb4b6cffee45014e72d4ddd6db8e78891dab3997

*tools/vendor/google3/blaze updated to c8038380f3ce81bdb639bd0f9a36ce2db93ec500
Move desugarLibraryConfigFiles to the module system

See Icb4b6cffee45014e72d4ddd6db8e78891dab3997

Bug: 262468597
Test: No semantic change for existing tests.
Change-Id: I3cc5f2ad04e13de075820e738ce804c666809885