Autoupdate of 2 submodules
*tools/adt/idea updated to 373f2e2523bb3c5caee72ac38eae8165c32b171a
Open logcat panel in debug mode

Automatically open logcat panel when the `Miscellaneous > Show logcat
automatically` option is enabled and the `General > Activate tool
window` is disabled.

This is a change of behaviour and not really a bugfix as it seems like
this was intended behaviour before:
cf https://b.corp.google.com/issues/267530392#comment2

Note: after testing manually from Android Studio, if the `Activate tool
window` is enabled, this will take precedence over the `Show logcat
automatically` option despite the logic here:
`run/DefaultStudioProgramRunner.kt?l=114`

Fixes: 267530392
Test: manual - followed repro steps from b/267530392

Change-Id: Ib93b93bcb6a542f7c4550caa59850f4ed8c20d48

*tools/vendor/google3/blaze updated to 7211624cbc51ad440d5d576d37e0ab728306da1e
[Blaze] Open logcat panel in debug mode

Automatically open logcat panel when the `Show logcat
automatically` option is enabled and the `Activate tool
window` is disabled.

This is a change of behaviour and not really a bugfix as it seems like
this was intended behaviour before:
cf https://b.corp.google.com/issues/267530392#comment2

Fixes: 267530392
Test: manual - followed repro steps from b/267530392 but using a Blaze
configuration instead. I used a cloudtop instance and the URL
go/codelab-aswb#importing-your-project in order to test this. (thanks
for the link Polina!)

Change-Id: I6cd86adc6f74e5839719dee7e2789e72e4db1f47