Autoupdate of 2 submodules
*tools/adt/idea updated to ec7982121d7d82be0e75bd304e17b123a90df974
adblib: Add "boost" support for JDWP processes

Given the `WAIT` packet behavior (see b/271572555 for
context) and the goal to eventually allow multiple instances
of Android Studio (or IntelliJ + Android plugin) running
concurrently when acquiring JDWP sessions on Android
device, this change implements "boosting" JDWP process
discovery when the Android Studio application is activated,
and "un-boosting" when de-activated.

This ensures the currently active Android Studio instance
will always have priority accessing the JDWP session
of new processes, which is critical for debugging (and
useful for other use cases, such are running an app).

The goal is to make sure we get to a state where having
multiple instances of Studio (or IJ+Android plugin)
running at the same time "just works" (*) from an
end-user perspective.

(*) "just works": most of the time, given that we
can't rely build a fool proof solution given the current
limitations of the Android Debugger and the way JDWP
sessions are contrained.

Bug: 271572555
Test: Included
Change-Id: I899bd8dafd0f5fb6e9c6e883aad7bc2aae5d7564

*tools/base updated to 7198741c2de4524827df7eb74768dc50e7f5551b
adblib: Improve behavior for WAIT packet

A previous change (I399ec064ab084ff4da1ba9cc07336515e0813522)
was made to keep the JDWP session open for a as long
as the timeout used by the process properties collector
(15 seconds currently).

However, given that close a JDWP session in the `WAIT` state
leaves the session completely unusable until the process is
terminated, this change updates the behavior to keep the
JDWP session open until either 1) the process is terminated
or 2) an external debugger (e.g. Android Studio) attaches
to the process. There is no timeout involved.

Bug: 271466829
Test: Included
Change-Id: I290cac4a107553f395e58a031ed2ecf03b8ee209