Sort filter result links lexicographically
We want links sorted so that presentation is more consistent across
refreshes. Lexicographic on (filepath, line, column) is easiest to do.

Implemented via a comparable data class generated via a functional
getter on LocationProvider, which is then used in VerifyActionMessage
via a `.sortedBy` in message generation.

We definitely need to clean up LocationProvider later between the code
duplication and read asserts.

Fixes: 272114763
Test: N/A
Change-Id: Id5f928513acde12183d86faf07da87152720af54