Add recitation warning UI (not wired up!)
This is the first part of the work. The UI is now
functional and able to display warnings for code
recitations, but the information needs to come
from the models. For now, you won't ever see the
warning IRL, as the infrastructure is not there
yet in the models' structure to provide such info.

Includes moving UserAvatar out of BotController.kt

Bug: 270663053
Test: manually tested
Change-Id: I35c14095b15e836516a1c8d0dbb780183267f19b