[toml] Tokenize square brackets more properly
When within an inline table, tokenize square brackets separately: the
only context in which two brackets makes a token is at toplevel (when
headers of tables and array-tables are acceptable).

This reveals some more undesireable resynchronizing behaviours, though
nothing that causes a crash; if we detect a problem when parsing a
key, restore the current location to be the start of the immediate
next token, rather than however far the parser had got, in order to
give consistent and predictable behavior.  Also, explicitly catch
unmatched close tokens and warn (if validating) rather than dropping
through to generic key/value parsing.

In the process, fix the start location of an invalid array of table to
be the first square bracket, not the second, by subtracting 2 from the
current offset after tokenizing rather than 1.

Bug: 272025574
Test: added
Change-Id: If7b9d731937f5dcacc7af6f495f829a0cbc376c4