Autoupdate of 2 submodules
*tools/adt/idea updated to 35053e90e0f9d4109136d1680d35f3fcdcdb2b12
Encapsulate Configuration of Monitor in a Class

Instead of passing a bunch of separate values.

This will come in useful when the Adblib based monitor is added.

Bug: 270737980
Test: Updated
Change-Id: I06393be3382208b7eca2a06714562d2ab06bd853

*tools/base updated to 4356015c4d7b1ebd55a8ecd1e4ff1f369b27f85a
Encapsulate Configuration of Monitor in a Class

Instead of passing a bunch of separate values.

This will come in useful when the Adblib based monitor is added.

Bug: 270737980
Test: Updated
Change-Id: Ic922edfe6929850ff45782058cbae3f7737ea02d