Run filters under progress indicators
Following the example from IntelliJ's TestCoroutineProgressAction,
this should simply be:

withBackgroundProgressIndicator(project, "Running...") {
  result.complete(smartReadAction(project) {
    filter.relevantLocations(project)
  })
}

However, deep inside the Psi reference search, PsiSearchHelper calls
CoreProgressManager#assertUnderProgress to check the progress
indicator in the current thread context. It doesn't work with the
coroutine context, which contains a wrapper (ProgressSink).
I didn't find a way to pass the original progress indicator through
all the contexts, so I used the classic non-blocking ReadAction.

Separately, when adding a test, I noticed AndroidCoroutineScope
handles exceptions by logging them as errors. While this is enough
when running in the IDE, in tests this meant assertion errors are
simply ignored and tests pass anyway. I added a custom logger so that
we can assert on these exceptions-as-errors, and made it an
ExternalResource in order to run automatically on all tests.

Bug: 263441269
Test: FilterServiceTest
Change-Id: Icacfa248b3742da1e7eb7ed2deaeae68c3e88162