Autoupdate of 2 submodules
*tools/adt/idea updated to 3038a7359c56a8d4c31402f094433ef0b8a76a94
Allow for runtime dependencies to be null

This change allows the extension of the ModelBuilderParameter to enable
the IDE to chose whether or not the variant dependencies model should
contain the runtime classpath or not on a per project basis.

This change currently maintains the old behaviour in all cases, it just
ensures the parameter is set and that the return type for the runtime
classpath is not nullable.

Bug: N/A
Test: existing snapshots should cover
Change-Id: I46a3ab563ef3e1afe9cb8cfab726de86eb6f23b8

*tools/base updated to 82b940824addbeb1b0fc5b2441ff7c331e06bfdb
Add controls for the IDE to decide whether to build the runtimeclasspath

This change adds a new method to the ModelBuilderParameter allowing
the IDE to prevent the model builder from building the runtime classpath.

Since AGP can't be opened with an older version of studio there should
be no case where this paremter exists but is not set. The reverse will
also be safe since reguardless to whether this is set the ModelBuilder
won't read it and will always produce the full classpath. This is
handled by Gradle and its use of proxys.

Bug: 267066187
Test: added test model snapshots
Change-Id: I5479d25e500510378093fafc50f6bc96be1a8cf3