Autoupdate of 2 submodules
*prebuilts/r8 updated to 3f2a7989fab00b7cb2417defd450d65a38fc6ecf
Update D8 R8 to 8.1.22-dev

This is a development snapshot, it's fine to use for studio canary
build, but not for BETA or release, for those we would need a release
version of R8 binaries. This build IS suitable for preview release
but IS NOT suitable for public release.

Built here: go/r8-releases/raw/8.1.22-dev
Test: ./gradlew check
Bug:

Change-Id: Idfb85f1da83e3a667798ac8e27caa0023be98a8f

*tools/base updated to 88dd532ab4505353c6797ca3b0ac3752ea9351c2
Update test expectation to account for R8 pruning of art profiles

Bug:
Change-Id: I76525d19a4cc6108555743ef648d194ff35a41f1
Test: tools/base/bazel/bazel test //tools/base/build-system/integration-test/application:library-tests --test_filter=ArtProfileMultipleLibrariesTest