Autoupdate of 2 submodules
*tools/adt/idea updated to a14cedb4ed40274634375f043c31289d1fa879d4
Use material3 theme for the Empty Views Activity.

This CL moves the generation code for the themes.xml for the values-v29
variant.
The themes is only used from the Basic Views Activity at the moment and
other activitie don't need it because Basic Views Activity attaches the
toolbar in the Activity's code..

Bug: 267714227
Test: Manually created Empty Views Activity, then Basic Views Activity
Change-Id: I519613bab28d466d2133e729cd0fd8f0e74eb7a6

*tools/base updated to 82a3d67b9d5325e40edd8315e64d94003b224466
Use material3 theme for the Empty Views Activity.

This CL changes the default theme for "Empty Views Activity" with a
material3 theme.

The issue happened because "Empty Views Activity" was created with a
material2 theme, then "Basic Views Activity" was created with a
material3 theme.

I think we can use material 3 themes for other activities, too by now.
Need to address it in a separate issue.

This CL also sorts out the "Basic Views Activity" specific theme creation
step.
Transparente system bars is only needed for "Basic Views Activity" and
the minimum required API for the following attributes is 23.
android:navigationBarColor
android:statusBarColor
android:windowLightStatusBar

Thus changed the location to declare the theme to make the system bars
as transparent within the recipe in the Basic Views Activity and change
the minimum required API as 23.

Bug: 267714227
Test: Manually created Empty Views Activity, then Basic Views Activity
Change-Id: Ibfb4a2d99dccb81db99c0163fc6b199ee07e46d0