Launch and wait for all filters at bundle creation
This ensures that problems will be prepopulated. The UI is a bit lacking
(the assistant window is just empty w/ a spinner), but there are
progress bars for running filters in the background task tray.

I'm a bit concerned that this could somehow run on the UI thread, but it
seems to run on worker threads in practice. I added @WorkerThread just
in case.

I tried to test this via mocking and verifying interactions, but our
version of Mockito doesn't play nice with suspended functions (like
`join` on a Job). I'd rather not make a super heavyweight test for this
right now so I'll come back to it later. We're still covered for the
purposes of "doesn't crash" due to other tests that create a bundle. We
just don't have a test enforcing the launch + wait behavior.

Fixes: 272576633
Test: N/A
Change-Id: Ic12cb61e8a77c03e9588a56279463f20d6100bf9