Fix no open file/language mismatch errors
Some errors arise when no file is open in the editor, or if there is a language mismatch between the editor file and the code response

If no file is open or there is a language mismatch, analysis is done in a fake context.
If no file is open, all actions are disabled.
If there is a language mismatch move actions are disabled.

Change-Id: I828d335c8197821bc9e22c60e612bae0e0e7b0d6