Display user avatar in IntelliJ IDEA as well
This required changing how the auth flow happens
in IJ, because we always need to obtain the user
picture, and in turn that always requires us to
exchange the authorization code for the OAuth 2.0
tokens.

There some general code and logic cleanup, but it
is limited as much as possible to the impacted
code to make the feature work. I have added a sign
out action as well, mirroring what's possible on
the AS side; and a way to know when the signed in
status changes, which the UI will have to listen
to in future CLs.

Bug: 272019814
Test: manually tested
Change-Id: I7057c319669e801f5a35c356fdbb4a584ee87f68