ui: Update the sidebar "Documentation" link
Update the sidebar "Documentation" link to point to the Perfetto documentation: https://perfetto.dev/docs

Linking to https://perfetto.dev/ requires the user to then find the "Doc" link in the upper right. This change skips an extra click and sends the user directly to the documentation they (presumably) want.

Change-Id: Ic2fb94eb6cdf810766121511351ebcfcab3acc14