Autoupdate of 3 submodules
*tools/adt/idea updated to 105de500967dbe7172d78264a65f6411e71699ea
Appropriately handle AndroidSyncException with sync issue reporting

When a gradle sync fail with a fatal error that result in no computed variant, ProjectImportAction fails with AndroidSyncException. This change adds the sync issues to the exception and report them through SyncIssueReporter. These issues are added to the DataNode and reported to the users along with the thrown exception. Throwing the sync exception fails the requested sync and the build output window is opened automatically to show users the sync issues with a link to the build file where the error originated.

Bug: 267726572
Test: Existing Tests
Change-Id: Ia0d7175893bbbb0c64bfd87362056cd45db0a1fc

*tools/analytics-library updated to b7ef95baca29023b526c0f23bc3336d1f484297b
Update studio_stats.proto with emty dimension sync error

Logs-Approval: cl/515674162
Bug: 267726572
Test: EmptyDimensionSyncErrorTest
Change-Id: I5073f2c9b004ab7cc7d5ff0eeac85282d9251828

*tools/base updated to 9b1e90f67b24ece35ad3169946e28d567131d0ae
Use new gradle sync issue on empty flavor dimension

Bug: 267726572
Test: EmptyDimensionSyncErrorTest
Change-Id: Iea78f729c93d521fe0bdf8c520137b0992c69390