Cache transform fixes, and only show intentions if a fix is available
Behavior details: Transform autofixes, implemented as quickfixes, will now only be available as intentions if a fix has been generated and is available to be applied immediately. If a change happens to the scope of the fix, the fix will be unregistered and regenerated. When the corresponding error is fixed or removed, the fix is likewise unregistered.

We still want to understand whether waiting for AIDA during a highlighting pass delays highlighting or affects the responsiveness of the editor.

Fixes: 272788035, 272788024
Change-Id: If0a300b9678fb20fbb35e108ba1d6b6d2e1ed3c4