Autoupdate of 2 submodules
*tools/adt/idea updated to 6d77a28110a383e9f3ccd656b9e8d371dfa356d4
Include full moduleName in Connection

Add ConnectionInferrer interface, which is implemented by Crashlytics.
Currently it contains code that is hard to test. With this change, tests
can easily fake it.

Test: n/a
Bug: 263259654
Change-Id: I93b968958b1df34bcecf6dce7b97f522afe5e067

*tools/vendor/google updated to 42b78a5c9fe0c6462d411d82d6ff5ebcced8d38d
Decouple firebase app from ConnectionsChanged

This change attempts to make ConnectionsChanged event not rely on
FirebaseApp which is specific to Crashlytics, but become more generic
so it can be used for vitals.

It seems the Module field was added to the FirebaseApp object entirely
to facilitate the inferring of new active connections. It is moved to
VariantConnection.

Test: added new test. Some parts are still untestable.
Bug: 263259654
Change-Id: I8f03e2eec40a785459f6e0d2bdebf16ca64e5977