Add metrics for code completion
This CL adds metrics for code completion: Events are sent at
4 specific times: when a request is first sent, when a response
is received, when a completion is shown and when it is accepted.

The trace id from the response is threaded through to all
the events except the request so that it is possible to tie
them together.

Currently, we only show the first completion option, so the
suggested completion index is always set to 0.

Metrics are sent using clearcut, but this is only active when
using Android Studio. With IntelliJ, we just log the metrics
events to idea.log.

Fixes: 272780928
Test: manual, verified via metrics tool window after a completion
Change-Id: I30479ce3c0c4582b8340d599dd13204313b47fe4