Fix Run tool tab being renamed to "Profile"
This CL introduces an ExecutionListener that restores the "Run" tool
window's stripe title after running a profiler executor.

The Run tool's window stripe title gets overridden using the executor's
actionName in ExecutionManagerImpl, both
when a process start is scheduled and when the process starts.

I attempted another approach in ag/21911689 and ag/21911690 which led to
a regression elsewhere. I also considered an approach of registering a
custom Run ToolWindow that would not allow overriding the stripe title
and icon. That approach did not work as it's not possible to register a
custom ToolWindow implementation.

The approach in this CL should be low risk as the listener is only
included in the profiler plugin and should only run after profiler
executors. Furthermore, the action undertaken (restoring the Run window
title and icon) should be a no-op as we expect that window to have that
name and icon.

Thanks, @kovalp, for the suggestion on this approach.

Fixes: 185167600
Test: new test + manual
Change-Id: I63deefd7c43c29a175af25eeadaa8d50001628ea