Don't register the change listener more than once per document
Now we maintain a counter for document that keeps track of the number of active fixes within that document. As long as this number is > 0, the change listener should be active in the document. When this number would hit zero, the listener is removed, and the next fix will re-register it.

Fixes: 273621478
Change-Id: I4ce7f1e9e127d738482df8d6a97d6700e01a3bb3