Use MaterialVdIconsProvider for autocomplete icons
This provider has logic to download up-to-date material icons and cache
them on disk. It's currently used for Asset Studio, and is easily reused
for this scenario.

I'm keeping the existing logic to load icons from resources as a
fallback. Since the downloaded icons change to reflect up-to-date
versions of the icons, the set that's included may differ slightly from
whatever version of the material library the user is developing against.
This fallback gives a chance to render icons that may have been removed
in a later version of the library.

Fixes: 272127130
Test: Included
Change-Id: Ib2637a478aba9630ebaa92782b93092e2851f88b