Pass JNA path to game tools
In order to make native calls, java must have a path to the jni dispatch
library. In order for this to work, the path needs to be set via
-Djna.boot.library.path and the library needs to be present.

If the path is not set, Java will default to a com/sun/... path. If the
library cannot be found at the given path, it will look through its
resources (e.g. jar files).

This path was not being set in the IntelliJ config (Game_Tools.xml)
causing it to fail at start up. This sets the path to the library in the
prebuilts directory.

Bug: 263887432
Test: Run game tools from within IntelliJ.
Change-Id: Ib461d586076dbeedd30440b0d0ededfb616fdd78