Pass JNA path to game tools
In order to make native calls, java must have a path to the jni dispatch
library. In order for this to work, the path needs to be set via
-Djna.boot.library.path and the library needs to be present.

If the path is not set, Java will default to a com/sun/... path. If the
library cannot be found at the given path, it will look through its
resources (e.g. jar files).

The start-up script used by AGDE and by the game-tools tests was not
properly setting this poperty nor was the game-tools packager including
the library.

This change updates the packager to include the library and updates the
two start-ip scripts (windows and linux) to set the jna path.

Bug: 263887432
Change-Id: If1a2030c82cfba9f105330a93490619da6208aa3