Replace deprecated "new project" function
IntelliJ deprecated the "new project" function we were using. In fact,
part of the old function would assert that we were on the main
dispatcher, which is not true when executing an RPC request in game
tools' server mode.

This change replaces that function call with the newer function. This
function does not care what thread we are on, resolving the failure we
were seeing.

Bug: 265953540
Change-Id: Icea966d632bee59bfb6bacdf7ba9497e53fdf171