Specify message bundle to use when instantiating actions
sml-ij-ce module includes its own messages bundle. As a result,
when loading the actions registered in sml-core, IntelliJ ends
up using the wrong message bundle and is unable to locate the
action text for actions registered in sml-core.xml. This CL fixes
the issue by explicitly specifying the bundle to use for actions.

Fixes: 274147933
Test: launch IJ plugin, verify that paste action can be invoked
Change-Id: I55c493f73c1bb48aad35b5f717f41c0fe99149f1