Dynamically set SKIP_GRADLE_TASKS_LIST default based on IDE
This ensures consistency with how we add omit_all_tasks in IDEA. This
prevents the notice on the Gradle window about tasks to not be displayed
in IDEA by default. In IDEA tasks are always present as we patch the
platform to disbable this.

See - https://git.jetbrains.org/?p=idea/android.git;a=commit;h=ea96580e73418b097d1b0670e2da7f8530aec248
Bug: N/A
Test: N/A
Change-Id: I34b79e1a1f40c4d77d3883e6c8db654d1dc4ecee