Always refresh after a build
This solves the case where a Fast Preview has failed to render and the
user presses the refresh button. In that case, even though there are not
changes in the files, we want to fully refresh the preview.

The optimization is probably not worth the edge cases, so we can always
refresh after a successful build.

Fixes: 274824383
Test: Manually verified
Change-Id: I024843b4fe24092f5643a7b406d6bad2ac704c0c