Switch to using Sources/JavaDocs/Samples from V2 libraries
If we have an Android Gradle Plugin version that populates these three
fields in the library models, use them to set up the IDE.

A flag has been added to control this and it should work for any builds
that include Ifd3e99786b48fc4e96a592f054051668a95bcb1b.

For versions of AGP before this change we leave the current approach of
using the injected model builder as is. The injected model builder
should not cost anything as long as we don't query the model.

Bug: 197636221
Test: snapshots
Change-Id: I2a6710e40748b948eb653d16605c7da56dc60edc