Autoupdate of 2 submodules
*prebuilts/tools updated to 67b1573da0d56ed55f459e93b5c3f1981249f672
Update Streaming Benchmarker app with version using new JDK

Bug: None
Test: None
Change-Id: Id12abe5d3c63e9e21b1e93c4d2a63a3473fe1f7d

*tools/adt/idea updated to 1f59c212f2c89a6f014e6a177ced487bc6a7eb40
Allow benchmarker to work consistently with emulators

Something changed recently with the way the emulator handles key
events in that some keys are not sent to the emulator until the key
is released. This change adds a KEY_RELEASED event to the KEY_PRESSED
event for dispatched keys so that they are properly received by the
emulator.

Benchmarking requires a sequence of numbers and other characters to
be entered by Studio, followed by the ENTER key. However, the ENTER
key was not being sent to emulators due to this issue.

Fixes: 267504025
Test: DeviceAdapterTest.kt
Change-Id: Id72ef991edb85bad30cb57cd753337f5fbb7d168