Use built-in html escaper
Before converting markdown to html, we do our own processing
of the markdown to escape html. However, commonmark's renderer
includes the capability to escape html, so this CL removes our
custom rendering in favor of commonmarks'.

Our html escaper wasn't correctly handling inline code blocks,
and we get a fix for that for free with this change.

Fixes: 275173453
Test: added
Change-Id: I21ab02faae44172b74f2f88569a3fa35e2f94e59