Hook up thumbs up/down actions to send metrics
This CL hooks up the thumbs up/down actions to send metrics via
the usage reporter.

The feedback actions were originally constructed as soon as the
user request was sent. However, this does not allow us to
customize the actions because at this time, we don't know the rpc
id. So this CL moves the construction of the feedback panel to
when the final end of message signal is received.

In addition, we also don't show the feedback panel for messages
that are restored from history: this is because those messages
don't retain their rpc id, but even more importantly, the rpc
ids may not really be valid after a few days.

There's some follow up work to be done in the UI: the thumbs up/
down icons don't have any state. As a result, it is impossible
to see if you've already provided feedback on a particular
message. Ideally, these would be stateful toggle buttons (where
toggling thumbs up automatically toggles off thumbs down).

Bug: 273964866
Test: manual, verified in idea.log and feedback panel
Change-Id: I00e2f41383653a781a7f05a513b4a3debaa34e0d