Parse code as a .kts file instead of a code block when finding errors
In the Kotlin parser we look for errors in the code. Currently we mistakenly parse the code as
a PsiBlockCodeFragment, where import statements are not valid and turned into error elements.
This causes false positives when code contains imports - the visible result is that move
actions are not made available.

Fixes: 275375026
Test: included
Change-Id: I17632f54ce9f2cfb42b2dc2a1aecac962fb10a6b