Autoupdate of 2 submodules
*prebuilts/tools updated to 2980b01dc4b8a731c4967593344b1be888779a98
Add upgraded compose dependency prebuilts

The compose module needs a higher version of the compose-desktop
and skiko-awt dependencies to address the lack of features compose
v1.1.1 has and the bugs skiko-awt v0.7.16 contains. The versions
needed are v1.3.1 (compose-desktop) and v0.7.55 (skiko-awt). In
order to fulfill these dependency needs, the v1.3.1 and v0.7.55
prebuilts of the needed dependencies have been added.

Bug: 272822954
Test: none
Change-Id: I8d579e112f58bbbcb980b1bd23ca1b5c92582474

*tools/adt/idea updated to 6b4b654f6d94f84bee15511946f9e14c6a554118
Upgrade/clean compose dependencies

The current prebuilt compose dependencies the adt-ui-compose
and compose-samples modules use are v1.1.1 for the compose-desktop
deps and v0.7.16 for skiko-awt. These older prebuilts not only
lack more modern features, but also have bugs that are fixed
in later versions. The compose-desktop deps have been upgraded
to v1.3.1 which is last stable release as of today. The skiko-awt
module has been upgraded to v0.7.55 which addresses a major bug
with swing-compose interop.

It was also noticed that there were two compose-desktop-foundation
modules libararies defined in adt-ui-compose module, but only the
latter one declared was needed. Thus, the first one was removed.

Bug: 272822954
Test: manual
Change-Id: Iae7f70b216d56ec9fbfc471e1679f68cf4e9ab51