LE: Avoid querying an Exception for a file name
When the compiler invocation fails with an exception, the exception object
might or might not contain a reference to the name of the file that caused
the failure.

In any case, we should not depend on that exception object. Instead,
we should just recompile all the files that failed in the next edit.

Bug: 275238534
Test: none
Change-Id: I775b305db8128d18474fca31205ac669fee35b1f