Normalize path returned by the emulator "avd path" command
https://r.android.com/2441481 added new behaviour that adds avd/../avd
to the path when the emulator returns AVD path information via
console commands.

Wear pairing uses this AVD path as a unique ID that needs to match
the initial AVD path. The new addition breaks this relation and makes
pairing fail.
The change normalizes the path to remove the .. from the path.

Fixes: 275128556
Test: EndToEndIntegrationTest
Change-Id: I95c2caf71d625d271106278f6255ada32f96fbe4