sml: Remove only the first <p>
When rendering markdown, we were removing all paragraphs.
This used to mostly work since at a lower level, we were
splitting up text by \n\n and returning them as multiple
chunks. However, it doesn't work for cases like \r\n.

The only "fix" we need is that we need to trim the first
<p> tag in an output since the markdown converter always
starts with a <p>, but our JLabel's (unlike a browser)
will actually render an empty line for those tags making
every question and response begin with an empty line.

So this CL removes the logic that removed all <p> tags,
and replaces it with a mardown postprocessor that inlines
the contents of the first <p> tag.

Fixes: 275567818
Test: added
Change-Id: I936bf3d07fc6810280ec52865af82610007648b2