Use common service pattern in SDK Manager
The SDK mananger works differently than the other game tools services.
When it opens, it will block the main thread. This makes it harder to
work with. This CL makes some changes to how we struct it, making it
possible to build it ontop of the common service and common window
wrapper.

Change-Id: I6d27b54b6f9d424413e618776c562f64c14f042c