Autoupdate of 2 submodules
*tools/base updated to 55c4ed727c7cba86df49c1bd09f8cc43897b7648
Add DeviceHandle.sourceTemplate

We need to know which template a device came from so that the
device manager can hide the template when the device is activated.

Also, clarify that DeviceTemplates should be immutable, since the
DeviceHandle's source template should not change underneath it.

Bug: 276454402
Test: n/a
Change-Id: I3877fedb536979d930884be94ba6b42e2333de5a

*tools/studio/google/cloud/testing updated to 79d1a75cb77e413910b565997445e039c76a8fd4
Add DeviceHandle.sourceTemplate

We need to know which template a device came from so that the
device manager can hide the template when the device is activated.

Bug: 276454402
Test: Updated
Change-Id: I33dc6f4aacb2e3936b2b4e9354b760de316c809f