Possibly change modelUpdater when updating Compose preview
We may want to change the modelUpdater of a given NlModel when updating
the Compose preview. For example, if we are in standard preview and
using the accessibility based model updater, when going to interactive
preview, we want to reuse the NlModel but not the accessibility updater
for performance reasons.

Bug: N/A
Test: N/A
Change-Id: I38921ec14fb1d41ab72bc3ab93e83722ac6406b3