Add missing imports for all of AndroidX
If a reference in a code response snippet is unresolved but is in AndroidX, we now add the right import to the model backing the snippet

Includes names of classes, interfaces, annotations, objects, and top-level functions from https://developer.android.com/reference/kotlin/androidx/packages.

Some common/simple names have been commented out, and duplicates are not considered.

Also shuffles the parser tests around (and renames one parser) so that testing the import feature is confined to
the Kotlin parser test and easier to find

Bug: 274832437
Change-Id: Id904552ec810f78dacf18a798d0e7649babb5bbe