[KTIJ-25101] Fix an AlreadyDisposedException
The module for a given Kotlin script is not necessarily the same
throughout the lifetime of a project; a sync might destroy and
re-create a module, meaning that we must recompute the module
each time we compute the target platform (rather than re-using
the module we initially computed).

Move the call to getScriptModule to close to its use, making it
clear that one is immediately-used and the other is at an
arbitrarily later time.

closes https://github.com/JetBrains/intellij-community/pull/2367

^KTIJ-25101 fixed

(cherry picked from commit 6282c7f059b3455eaa79e5eb4ee87c58a194c796)

Change-Id: I6282c7f059b3455eaa79e5eb4ee87c58a194c796