Autoupdate of 2 submodules
*tools/adt/idea updated to ef5b08da39fd6969101a9026eb4046a112285354
Move foldable AVD config to xml

Currently half of the foldable specific configs are in java. Let's move
them into xml so that they are in one place.

Bug: 275861444
Test: AvdManagerConnectionTest
Change-Id: I605370b92a947ddf33481e0049d60846247e4571

*tools/base updated to 4494a4329766b361d81123ff8e4fb87f33a9ab8e
Move foldable AVD config to xml

Currently half of the foldable specific configs are in java. Let's move
them into xml so that they are in one place.

This commit added hinge config for foldable deviceis. It created
devices-7.xsd for the hinge config.

Bug: 275861444
Test: AvdManagerConnectionTest
Change-Id: Ia5fc3a69cb744fd43c62c5c4c2c3dbfb2f460d86