Fix avatars not being updated as needed
AvatarProvider is no more, now we have a central
place that listens for avatar changes, and keeps
a cache: UserAvatarService. If the avatar is not
immediately ready when the user logs in (as it
is not in Studio's auth plugin), it starts polling
until a new avatar is available, for up to 1 min.

The centralisation allows us to conserve resources
as much as possible, exposing a shared flow that
UI components can listen to, getting all the
benefits of this implementation without effort nor
significant runtime cost.

ProfilePicture now listens for avatar changes (not
user auth changes, as it was before) and fetches
the up-to-date avatar as it's available.

This CL also tweaks/fixes a few things:
 * Onboarding steps would throw on init
 * In some cases, showing/hiding the timeline's
   empty state would throw
 * Add a UI Inspector provider to some components,
   to aid the debugging. This has no runtime cost
   if the Inspector isn't open.
 * Tweak logging severity to reduce spamming

Bug: 275245686
Test: manually tested
Change-Id: I26cda03421724c5e544a9a6ea0073132ffd0d3e4