Only qualify simple name references at the start of a qualified expression
Previously the qualification visitor ignored all references within qualified expressions. We should still qualify such references if and only if they are simple name references and they are at the start of the qualified expression.

There was one exception - if references in a qualified expression had names covered by imports included in the snippet, they could be mistakenly qualified. This bug is also fixed.

Test: included
Change-Id: If2f2f7b456c57e8fd213dbc831fdf64a8845fe9a