Remove "extract function" dialog
Instead of prompting the user to choose the name of a function to be created from a code snippet consisting of statements, we now create a function with a hardcoded name and place it in the file automatically. The user still gets a choice of scope, but no additional dialogs.

Also enables two tests which have been disabled for a long time because of dependence on this dialog

Fixes: 276977441
Test: included
Change-Id: I0dee6e23bef842905f754dc2b1d2aca5a3d62daa