Use constant for default emulator image
This change simply replaces hard-coded instances of the API-level-29
emulator with a reference to INTEGRATION_TEST_SYSTEM_IMAGE. As such,
there should be no functional change here; it will simply be easier to
update the emulator image in the future.

Note that LayoutEditorPreviewTest.kt didn't actually use an emulator at
all, so I removed it from the BUILD file.

Bug: 274505940
Test: N/A - no functional changes
Change-Id: Ida7752ed1560d310078c8245369e1f2a46a33852