Prepare UI to display verified messages
Given the absence of any specs on how a verified
message is transmitted, I have made up something,
assuming they are made up by chunks, and are not
streamed in but rather produced with the initial
message as "special" chunks.

A subsequent CL will need to rectify any incorrect
assumptions and wire things up to the actual data,
once it appears in the protos.

This CL focuses merely on the UI side of things.

Bug: 276916339
Test: manually tested
Change-Id: I7096f6f28758f88a41050bbd4fde17a2e9847bff